% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicModel.R
\name{aicModel}
\alias{aicModel}
\title{Variable selection using AIC.}
\usage{
aicModel(solution_path, coef_values, mean, x, y, t)
}
\arguments{
\item{solution_path}{Sequence of models produced by penalized credible regions method.
A \eqn{p \times p} matrix whose (i,j) element is TRUE
when the j-th parameter belongs into the i-th model else FALSE.}

\item{coef_values}{Values of the coefficients obtained by creating the model sequence.}

\item{mean}{Mean of the regression parameter at time t. A vector of length p.}

\item{x}{Covariance vector at time t. A vector of length t.}

\item{y}{Value of the response variable at time t.}

\item{t}{Time step for which the models are evaluated.}
}
\value{
The selected variables according to AIC. A vector containing the selected variables.
}
\description{
Select variables according to the AIC for a dynamic binary logistic model based
on a solution path obtained by applying the penalized credible regions method.
}
