% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlot.R
\name{calculateAUC}
\alias{calculateAUC}
\title{Calculate AUC of a ROC curve}
\usage{
calculateAUC(data)
}
\arguments{
\item{data}{data frame containing columns FPR and TPR that the ROC curve is based on}
}
\value{
AUC value
}
\description{
Calculate AUC of a ROC curve
}
