% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicModel.R
\name{dicModel}
\alias{dicModel}
\title{Variable selection using DIC.}
\usage{
dicModel(beta_chain, solution_path, mean, x, y, t, burn_in = 500)
}
\arguments{
\item{beta_chain}{A chain of values for the regression parameters of a DGLM produced by an MCMC algorithm.}

\item{solution_path}{Sequence of models produced by penalized credible regions method.
A \eqn{p \times p} matrix whose (i,j) element is TRUE
when the j-th parameter belongs into the i-th model else FALSE.}

\item{mean}{Mean of the regression parameter at time t. A vector of length p.}

\item{x}{Covariance vector at time t. A vector of length t.}

\item{y}{Value of the response variable at time t.}

\item{t}{Time step for which the models are evaluated.}

\item{burn_in}{The burn in period for the MCMC chain.}
}
\value{
The selected variables according to DIC. A vector containing the selected variables.
}
\description{
Select variables according to the DIC for a dynamic binary logistic model based
on a solution path obtained by applying the penalized credible regions method.
}
