% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFBS.R
\name{FFBS}
\alias{FFBS}
\title{Forward filtering backward sampling}
\usage{
FFBS(y, x, m0, C0, G, W, V)
}
\arguments{
\item{y}{Realizations of the response variables. A vector of length \eqn{T}.}

\item{x}{Covariate vectors. A \eqn{T \times p} matrix.}

\item{m0}{Prior mean of \eqn{\boldsymbol{\beta}_0}. A vector of length \eqn{p}.}

\item{C0}{Prior covariance of \eqn{\boldsymbol{\beta}_0}. A \eqn{p \times p} matrix.}

\item{G}{Evolution matrices. An \code{R} function \code{G(t)} that returns the \eqn{p \times p} evolution matrix \eqn{G_t} dependent on the time argument \eqn{t}.}

\item{W}{Error covariance. A \eqn{p \times p} matrix.}

\item{V}{Variances of the observation variables. A vector of length \eqn{T}.}
}
\value{
A sample from the posterior distribution \eqn{p(\boldsymbol{\beta}_{1:T} \: \vert \: \boldsymbol{y}_{1:T})} for the
DLM \eqn{(G_t, W, 1/V_t)_{t=1}^T}. A \eqn{(T+1) \times p} matrix whose \eqn{(t,i)} element
is the sampled value for \eqn{\beta_{t,i}}.
}
\description{
Applies the forward filtering backward sampling algorithm sample the from the posterior distribution of the regression parameters of a DLM.
Modification of the code \url{https://github.com/jordanbakerman/Dynamic/blob/main/FFBS_Function_Code.txt}.
}
