% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msESS.R
\name{msESS}
\alias{msESS}
\title{Monotone sequence estimator for ESS}
\usage{
msESS(chain)
}
\arguments{
\item{chain}{Vector containing values produced by an MCMC algorithm.}
}
\value{
The montone sequence estimator of the chain.
}
\description{
Monotone sequence estimator of the effective sample size (ESS) for an MCMC chain.
}
