% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlot.R
\name{rocPlot}
\alias{rocPlot}
\title{ROC plots and AUC of the solution paths.}
\usage{
rocPlot(solution_path, true_model, method)
}
\arguments{
\item{solution_path}{Sequence of models produced by penalized credible regions method.
A \eqn{p \times p} matrix whose (i,j) element is TRUE
when the j-th parameter belongs into the i-th model else FALSE.wise false.}

\item{true_model}{Vector of length p indicating which variables belong into the true model.
The i-th element is \code{TRUE} if the i-th element belongs into the true model else
it is \code{FALSE}.}

\item{method}{Name of the method used. Will be included into the title of the plot.}
}
\value{
A named list. First element \code{plot} is the ROC-curve plot and the second element
\code{AUC} is the AUC value for the ROC-curve.
}
\description{
Creates a ROC plot for a solution path and calculates its AUC.
}
