% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomDglmGibbs.R
\name{sampleLatentPg}
\alias{sampleLatentPg}
\title{Sample latent polya-gamma variables}
\usage{
sampleLatentPg(b, x, beta)
}
\arguments{
\item{b}{parameter of the polya-gamma distribution; vector of length T}

\item{x}{covariates; a Txp matrix}

\item{beta}{values of the regression parameters; (T)xp matrix}
}
\value{
T samples from a Polya-Gamma distribution, where the t-th sample is
from a PG(b\link{t}, x_t^Tbeta_t + offset) distribution
}
\description{
Sample latent polya-gamma variables
}
