% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomDglmGibbs.R
\name{updateError}
\alias{updateError}
\title{Updates the values of the rate parameter of the error covariance component.}
\usage{
updateError(gamma, beta, G)
}
\arguments{
\item{gamma}{Values of the gamma parameter. Vector of length p.}

\item{beta}{Values of the regression parameters; \eqn{T \times p} matrix.}

\item{G}{Evolution matrices. A function of time.}
}
\value{
Updated value of the gamma parameters. A vector of length p.
}
\description{
Updates the values of the rate parameter of the error covariance component.
}
