% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableRanking.R
\name{variableRanking}
\alias{variableRanking}
\title{Variable ranking table.}
\usage{
variableRanking(model_sequence)
}
\arguments{
\item{solution_path}{Sequence of models produced by penalized credible regions method.
A \eqn{p \times p} matrix whose (i,j) element is TRUE
when the j-th parameter belongs into the i-th model else FALSE.}
}
\value{
A data frame with rows \code{model_index} and \code{variable}.
The row \code{model_index} contains the index of the model in the sequence.
The row \code{variable} contains the variable (index).
}
\description{
Creates a variable ranking table based on a solution path. The table shows
for each model in the sequence which variable was newly included into this model.
}
