% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableSelectionTable.R
\name{variableSelectionTable}
\alias{variableSelectionTable}
\title{Variable selection.table}
\usage{
variableSelectionTable(selected_model, true_model, p, t)
}
\arguments{
\item{selected_model}{Model selected by some method. A vector containing the variables included into the model.}

\item{true_model}{The true model. A vector containing the variables included into the model.}

\item{p}{Number of parameters.}

\item{t}{Time step at which the model was selected}
}
\value{
Data frame containing the columns \code{p}, \code{T}, \code{TPR} and \code{FPR} with the respective values
for the selected model.
}
\description{
Creates a table to evaluate variable selection by calculating TPR and FPR values.
}
